requirements.txt
pillow
streamlit
openpyxl
